--- src/index.html
@@ -1,95 +1,0 @@
-<!DOCTYPE html>
-
-<!--
-    rdd - https://github.com/latte-soft/rdd
-
-    Copyright (C) 2024 Latte Softworks <latte.to> | MIT License
--->
-
-<html lang=”en”>
-    <head>
-        <title>rdd</title>
-        <link rel="icon" href="favicon.svg">
-
-        <style>
-            @import url('https://fonts.googleapis.com/css2?family=JetBrains+Mono&display=swap');
-
-            :root {
-                color-scheme: dark light;
-            }
-
-            body {
-                font-family: 'JetBrains Mono', monospace;
-            }
-
-            form {
-                border-width: thin;
-                border-style: solid;
-                border-radius: 4px;
-                margin-left: 10px;
-                width: fit-content;
-                padding: 18px 18px 18px 18px;
-                line-height: 24px;
-            }
-
-            output {
-                white-space: pre-wrap;
-            }
-
-            button {
-                font-weight: normal;
-            }
-
-            a:link {
-                color: rgb(43, 162, 241);
-                font-weight: bolder;
-            }
-
-            a:visited {
-                color: rgb(43, 162, 241);
-                font-weight: bolder;
-            }
-        </style>
-    </head>
-
-    <body>
-        <p>RDD - "Roblox Deployment Downloader" &#40;<a target="_blank" href="https://github.com/latte-soft/rdd">GitHub Repository</a>, <a target="_blank" href="https://latte.to/discord">Latte Softworks Discord Server</a>&#41;<br>Copyright (c) 2024 Latte Softworks <a target="_blank" href="https://latte.to">&lt;latte.to&gt;</a> | <a target="_blank" href="https://github.com/latte-soft/rdd/blob/master/LICENSE">MIT License</a></p>
-        <noscript><br>&#40;⚠️ This site doesn't work without JavaScript Enabled!&#41;<br></noscript>
-
-        <!-- Shown by rdd.js if there's no query args -->
-        <div hidden id="downloadFormDiv">
-            <form id="downloadForm" target="_blank" action="">
-                <label for="binaryType">Binary Type:</label>
-                <select id="binaryType" name="binaryType">
-                    <option value="WindowsPlayer">WindowsPlayer</option>
-                    <option value="WindowsStudio64">WindowsStudio64</option>
-                    <option value="MacPlayer">MacPlayer</option>
-                    <option value="MacStudio">MacStudio</option>
-                </select>
-                <br>
-                <label for="channel">Channel Name:</label>
-                <input id="channel" name="channel" type="text" placeholder="LIVE">
-                <br>
-                <label for="version">Version Hash:</label>
-                <input id="version" name="version" type="text"> <b>&#40;Optional&#41;</b>
-                <br><br>
-                <label for="compressZip">Compress Output Zip:</label>
-                <input id="compressZip" name="compressZip" type="checkbox">
-                <br>
-                <label for="compressionLevel">Zip Compression Level (1-9):</label>
-                <input id="compressionLevel" name="compressionLevel" type="number" min="1" max="9" value="5">
-                <br><br>
-                <button type="button" onclick="downloadFromForm();">Download</button>
-                <button type="button" onclick="copyLinkFromForm();">Copy Permanent Link</button>
-                <br>
-            </form>
-            <br>
-        </div>
-
-        <output id="consoleText"></output>
-        <br>
-    </body>
-
-    <script src="js/jszip.min.js"></script>
-    <script src="js/rdd.js"></script>
-</html>
