+++ tsconfig.app.tsbuildinfo
@@ -1,0 +1,1 @@
+{"root":["./src/rdd.ts"],"version":"5.6.2"}
